requirements.txt
Django==3.1
djangorestframework==3.11.1
python-dotenv>=0.9.1
requests==2.24.0
psycopg2-binary==2.8.5
django-filter==2.3.0
django-crispy-forms==1.9.2

#testing
black==19.10b0
pylint==2.5.3
pylint-django==2.3.0
pylint-plugin-utils==0.6